+++ src/dpsprep/concurrency/__init__.py
@@ -1,1 +1,2 @@
-from .api import concurrently_process_document
+from .api import concurrently_process_pages
+from .counter import PageCounter